Fixed the splits
Wow, found that the splits were not being calculated at all correctly. The seconds presented were the seconds from the overall time, since the seconds from the splits were computed wrong, as were the minutes.

We were subtracting minutes first then the seconds separately, but we were not correcting for times when say we are subtracting 8 minutes 20 seconds overall after the second KM from 4 minutes and 40 seconds for the first KM to get the 2nd KM split, we were gettings -20 seconds and 4 minutes, but that should be 3 minutes and 40 seconds, that is now fixed